- renames config commands to workspace
