+++ src/kiota/Handlers/Workspace/MigrateHandler.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 using Kiota.Builder.WorkspaceManagement;
 using Microsoft.Extensions.Logging;
 
-namespace kiota.Handlers.Config;
+namespace kiota.Handlers.Workspace;
 
 internal class MigrateHandler : BaseKiotaCommandHandler
 {
