+++ src/kiota/KiotaWorkspaceCommands.cs
@@ -1,13 +1,13 @@
 ﻿using System.CommandLine;
-using kiota.Handlers.Config;
+using kiota.Handlers.Workspace;
 
 namespace kiota;
 
-public static class KiotaConfigCommands
+public static class KiotaWorkspaceCommands
 {
-    public static Command GetConfigNodeCommand()
+    public static Command GetWorkspaceNodeCommand()
     {
-        var command = new Command("config", "Manages the Kiota configuration");
+        var command = new Command("workspace", "Manages the Kiota workspace configuration");
         command.AddCommand(GetInitCommand());
         command.AddCommand(GetMigrateCommand());
         return command;
@@ -15,7 +15,7 @@ public static class KiotaConfigCommands
     private static Command GetInitCommand()
     {
         var logLevelOption = KiotaHost.GetLogLevelOption();
-        var command = new Command("init", "Initializes the Kiota configuration"){
+        var command = new Command("init", "Initializes the Kiota workspace configuration"){
             logLevelOption,
         };
         command.Handler = new InitHandler
@@ -29,7 +29,7 @@ public static class KiotaConfigCommands
         var logLevelOption = KiotaHost.GetLogLevelOption();
         var lockDirectoryOption = GetLockDirectoryOption();
         var classOption = KiotaClientCommands.GetClientNameOption(false);
-        var command = new Command("migrate", "Migrates a kiota lock file to a Kiota configuration")
+        var command = new Command("migrate", "Migrates a kiota lock file to a Kiota workspace configuration")
         {
             logLevelOption,
             lockDirectoryOption,
